adjust pr based on previous discussion
