+++ docs/src/contributing/assertions.rst
@@ -1,11 +1,11 @@
-.. _assertions_verification:
+.. _assertions_check_arg:
 
-============================
-Assertions and Verification
-============================
+=================================
+Assertions and Argument Checking
+=================================
 
 LVGL provides two sets of macros for validating conditions at runtime:
-**assertions** and **verification**. This page explains when and how to use
+**assertions** and **argument checking**. This page explains when and how to use
 each within LVGL's codebase.
 
 
@@ -14,7 +14,7 @@ Overview
 ********
 
 +---------------------------+---------------------------+---------------------------+
-| Feature                   | LV_ASSERT                 | LV_VERIFY                 |
+| Feature                   | LV_ASSERT                 | LV_CHECK_ARG              |
 +===========================+===========================+===========================+
 | Purpose                   | Catch programming errors  | Validate runtime input    |
 +---------------------------+---------------------------+---------------------------+
@@ -30,7 +30,7 @@ Overview
 **Use assertions** for conditions that indicate bugs in the code—situations
 that should never happen if the code is correct.
 
-**Use verification** for conditions that might legitimately fail at runtime
+**Use argument checking** for conditions that might legitimately fail at runtime
 (e.g., user-provided input, resource availability) and require graceful handling.
 
 
@@ -107,14 +107,14 @@ Specialized Assertions
 
 
 
-Verification
-************
+Argument Checking
+*****************
 
-Verification macros provide a more flexible approach than assertions: they log
-a warning and allow you to specify a recovery action (such as returning from a
-function or continuing with a fallback).
+The :c:macro:`LV_CHECK_ARG` macro provides a flexible approach for validating
+function arguments at runtime: it logs a warning and allows you to specify a
+recovery action (such as returning from a function or continuing with a fallback).
 
-These macros are useful for:
+This macro is useful for:
 
 - Input validation at the start of functions
 - Checking preconditions without halting the program
@@ -125,14 +125,21 @@ These macros are useful for:
 Configuration
 -------------
 
-The verification macros use the :ref:`logging` system to output warnings.
-Make sure logging is enabled by setting :c:macro:`LV_USE_LOG` to ``1`` in ``lv_conf.h``.
+In ``lv_conf.h``:
+
+- :c:macro:`LV_USE_CHECK_ARG` - Set to ``1`` to enable, ``0`` to disable.
+  When disabled, all ``LV_CHECK_ARG`` calls compile to nothing.
+- :c:macro:`LV_CHECK_ARG_ASSERT_ON_FAIL` - Set to ``1`` to also call
+  ``LV_ASSERT_HANDLER`` on failure (before the action is executed).
+
+The macro uses the :ref:`logging` system to output warnings.
+Make sure logging is enabled by setting :c:macro:`LV_USE_LOG` to ``1``.
 
 
-Basic Verification
-------------------
+Usage
+-----
 
-:c:macro:`LV_VERIFY`
+:c:macro:`LV_CHECK_ARG`
    Takes a condition, an action to execute on failure, and optional printf-style
    arguments for the log message:
 
@@ -140,98 +147,32 @@ Basic Verification
 
       void draw_to_display(lv_obj_t * display, int len)
       {
-          /* Return early if obj is NULL */
-          LV_VERIFY(display != NULL, return, ": display must be provided - we can't draw onto thin air!");
+          /* Return early if display is NULL */
+          LV_CHECK_ARG(display != NULL, return, ": display must be provided");
+
+          /* Return -1 if len is invalid, logging the actual value */
+          LV_CHECK_ARG(len % 2 == 0, return, ": need even data points, but got %d", (int)len);
       }
 
-      int get_deviation(int** data, size_t len)
+      int get_width(lv_obj_t * obj)
       {
-          /* Return -1 if len is invalid, logging the actual value */
-          LV_VERIFY(len % 2 == 0, return -1, ": need an even number of data points, but got %d", (int)len);
-
-          /* Continue with normal operation... */
+          /* Return -1 if obj is NULL */
+          LV_CHECK_ARG(obj != NULL, return -1, "tried to get width of NULL object");
+          /* ... */
       }
 
    The condition is stringified and included in the log message, so you can see
    exactly what check failed. Only supply a format string and arguments if you
    want to include additional context beyond the condition itself.
 
-:c:macro:`LV_VERIFY_MSG`
-   Same as :c:macro:`LV_VERIFY` but with a plain string message:
+   Using ``break`` as the action in a loop:
 
    .. code-block:: c
 
-      const char * error_msg = get_validation_error();
-      LV_VERIFY_MSG(is_valid, return, error_msg);
-
-
-Log-Only Verification
----------------------
-
-Sometimes you want to log a warning without taking any action:
-
-:c:macro:`LV_VERIFY_OR_LOG`
-   Log with a printf-style format:
-
-   .. code-block:: c
-
-      /* Log if obj is NULL, but continue anyway */
-      LV_VERIFY_OR_LOG(screen != NULL, ", using primary screen (%s)", primary_screen_name);
-
-:c:macro:`LV_VERIFY_OR_LOG_MSG`
-   Log with a plain string message:
-
-   .. code-block:: c
-
-      LV_VERIFY_OR_LOG_MSG(obj != NULL, "continuing with defaults");
-
-
-Return Helpers
---------------
-
-For common return patterns, convenience macros are provided:
-
-:c:macro:`LV_VERIFY_OR_RETURN`
-   Return from a void function:
-
-   .. code-block:: c
-
-      void draw_to_display(lv_obj_t * display)
-      {
-          LV_VERIFY_OR_RETURN(display != NULL, "display needs to be provided");
+      for(int i = 0; i < count; i++) {
+          LV_CHECK_ARG(data[i] != NULL, break, ": NULL at index %d", i);
           /* ... */
       }
-
-:c:macro:`LV_VERIFY_OR_RETURN_VAL`
-   Return a specific value:
-
-   .. code-block:: c
-
-      int get_width(lv_obj_t * obj)
-      {
-          LV_VERIFY_OR_RETURN_VAL(obj != NULL, -1, "tried to get width of NULL object, returning -1");
-          /* ... */
-      }
-
-
-Assert on Failure
------------------
-
-When a condition is critical and should halt the program (be careful to have your assert handler configured appropriately for non-debug builds):
-
-:c:macro:`LV_VERIFY_OR_ASSERT`
-   Logs a warning and invokes :c:macro:`LV_ASSERT_HANDLER`:
-
-   .. code-block:: c
-
-      LV_VERIFY_OR_ASSERT(len % 2 == 0, "data must come in pairs, but got %d data points", len);
-
-:c:macro:`LV_VERIFY_OR_ASSERT_MSG`
-   Same as above, but with a plain string message:
-
-   .. code-block:: c
-
-      LV_VERIFY_OR_ASSERT_MSG(obj != NULL, "can't draw on thin air");
 
 
 
@@ -261,8 +202,8 @@ Assertions
      - Assert memory pool integrity (if enabled)
 
 
-Verification
-------------
+Argument Checking
+-----------------
 
 .. list-table::
    :widths: 35 65
@@ -270,20 +211,5 @@ Verification
 
    * - Macro
      - Description
-   * - :c:macro:`LV_VERIFY`
-     - Verify condition; on failure log and execute action
-   * - :c:macro:`LV_VERIFY_MSG`
-     - Same as LV_VERIFY but with plain string message
-   * - :c:macro:`LV_VERIFY_OR_LOG`
-     - Verify condition; on failure log only (no action)
-   * - :c:macro:`LV_VERIFY_OR_LOG_MSG`
-     - Same as LV_VERIFY_OR_LOG but with plain string message
-   * - :c:macro:`LV_VERIFY_OR_RETURN`
-     - Verify condition; on failure log and return (void function)
-   * - :c:macro:`LV_VERIFY_OR_RETURN_VAL`
-     - Verify condition; on failure log and return specified value
-   * - :c:macro:`LV_VERIFY_OR_ASSERT`
-     - Verify condition; on failure log and call LV_ASSERT_HANDLER
-   * - :c:macro:`LV_VERIFY_OR_ASSERT_MSG`
-     - Same as LV_VERIFY_OR_ASSERT but with plain string message
-
+   * - :c:macro:`LV_CHECK_ARG`
+     - Check condition; on failure log and execute action
