+++ lv_conf_template.h
@@ -517,6 +517,20 @@
 #define LV_ASSERT_HANDLER while(1);     /**< Halt by default */
 
 /*-------------
+ * Check arg
+ *-----------*/
+
+/** Enable LV_CHECK_ARG macro to validate function arguments at runtime.
+ * When enabled, failed checks log a warning and execute the specified action.
+ * 0: Disable all LV_CHECK_ARG checks (checks compile to nothing)
+ * 1: Enable LV_CHECK_ARG checks */
+#define LV_USE_CHECK_ARG 1
+
+/** If enabled, also call LV_ASSERT_HANDLER when an LV_CHECK_ARG check fails.
+ * Requires LV_USE_CHECK_ARG to be enabled. */
+#define LV_CHECK_ARG_ASSERT_ON_FAIL 0
+
+/*-------------
  * Debug
  *-----------*/
