+++ lvgl.h
@@ -36,7 +36,7 @@ extern "C" {
 #include "src/misc/lv_iter.h"
 #include "src/misc/lv_circle_buf.h"
 #include "src/misc/lv_tree.h"
-#include "src/misc/lv_verify.h"
+#include "src/misc/lv_check_arg.h"
 
 #include "src/osal/lv_os.h"
